Fix review comments in #82
