+++ presto-main/src/test/java/com/facebook/presto/ingest/TestStreamWriterTupleValueSink.java
@@ -1,10 +1,15 @@
 package com.facebook.presto.ingest;
 
-import com.facebook.presto.block.*;
+import com.facebook.presto.block.Blocks;
+import com.facebook.presto.block.Cursor;
+import com.facebook.presto.block.Cursors;
+import com.facebook.presto.block.QuerySession;
+import com.facebook.presto.block.TupleStream;
+import com.facebook.presto.block.TupleStreamSerde;
+import com.facebook.presto.block.TupleStreamSerdes;
 import com.facebook.presto.operator.GroupByOperator;
 import com.facebook.presto.slice.Slices;
 import com.google.common.io.OutputSupplier;
-import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import java.io.ByteArrayOutputStream;
@@ -13,18 +18,12 @@ import java.io.OutputStream;
 
 public class TestStreamWriterTupleValueSink
 {
-    private StreamWriterTupleValueSink streamWriterTupleValueSink;
-    private ByteArrayOutputStream byteArrayOutputStream;
-    private TupleStreamSerde serde;
-    private TupleStream tupleStream;
-    private Cursor cursor;
-
-    @BeforeMethod(alwaysRun = true)
-    public void setUp() throws Exception
+    @Test
+    public void testSanity() throws Exception
     {
-        byteArrayOutputStream = new ByteArrayOutputStream(1024);
-        serde = TupleStreamSerdes.Encoding.RAW.createSerde();
-        streamWriterTupleValueSink = new StreamWriterTupleValueSink(
+        final ByteArrayOutputStream byteArrayOutputStream = new ByteArrayOutputStream(1024);
+        TupleStreamSerde serde = TupleStreamSerdes.Encoding.RAW.createSerde();
+        StreamWriterTupleValueSink streamWriterTupleValueSink = new StreamWriterTupleValueSink(
                 new OutputSupplier<OutputStream>() {
                     @Override
                     public OutputStream getOutput() throws IOException
@@ -34,13 +33,9 @@ public class TestStreamWriterTupleValueSink
                 },
                 serde.createSerializer()
         );
-        tupleStream = new GroupByOperator(Blocks.createTupleStream(0, "a", "a", "b", "b"));
-        cursor = tupleStream.cursor(new QuerySession());
-    }
-
-    @Test
-    public void testSanity() throws Exception
-    {
+        TupleStream tupleStream = new GroupByOperator(Blocks.createTupleStream(0, "a", "a", "b", "b"));
+        Cursor cursor = tupleStream.cursor(new QuerySession());
+        
         cursor.advanceNextValue();
         streamWriterTupleValueSink.process(Cursors.asTupleStreamPosition(cursor));
         cursor.advanceNextValue();
